+++ apps/server/src/database/postgres.rs
@@ -1,8 +1,7 @@
-use sqlx::{postgres::PgPoolOptions, PgPool};
 use crate::config::Config;
+use sqlx::{PgPool, postgres::PgPoolOptions};
 
-
-pub async fn connect (config:&Config) -> Result<PgPool, sqlx::Error> {
+pub async fn connect(config: &Config) -> Result<PgPool, sqlx::Error> {
     PgPoolOptions::new()
         .max_connections(5)
         .connect(&config.database_url)